+++ testdata/webapp1/config/aah.conf
@@ -65,7 +65,7 @@ server {
 
     # aah server graceful shutdown timeout
     # Default value is `60s`.
-    #grace_shutdown = "60s"
+    grace_shutdown = "60h"
   }
 
   # Mapped to `http.Server.MaxHeaderBytes`.
