+++ testdata/webapp1/config/routes.conf
@@ -179,6 +179,7 @@ domains {
         controller = "testSiteController"
         method = "post"
         action = "FormSubmit"
+        anti_csrf_check = true
       }
 
       create_record {
@@ -250,6 +251,7 @@ domains {
           ws_text {
             path = "/text"
             method = "WS"
+            websocket = "testWebSocket"
             action = "Text"
           }
           ws_binary {
